refactor(transformer): extract Responses logger
