+++ open-sse/transformer/responsesTransformer.ts
@@ -1,76 +1,11 @@
-import { appendToolCallArgumentDelta } from "../utils/toolCallArguments.ts";
 import { shouldParseTextualReasoningTags } from "../handlers/responseSanitizer.ts";
-import * as fs from "fs";
-import * as path from "path";
+import { appendToolCallArgumentDelta } from "../utils/toolCallArguments.ts";
+export { createResponsesLogger } from "./responsesLogger.ts";
 /**
  * Responses API Transformer
  * Converts OpenAI Chat Completions SSE to Codex Responses API SSE format
  * Can be used in both Next.js and Cloudflare Workers
  */
-
-// Dynamic import for Node.js-only modules (fs/path unavailable in Workers)
-let _fs = null;
-let _path = null;
-async function getFs() {
-  if (_fs === null) {
-    try {
-      _fs = (await import("fs")).default;
-    } catch {
-      _fs = false;
-    }
-  }
-  return _fs || null;
-}
-async function getPath() {
-  if (_path === null) {
-    try {
-      _path = (await import("path")).default;
-    } catch {
-      _path = false;
-    }
-  }
-  return _path || null;
-}
-
-// Create log directory for responses (Node.js only)
-export function createResponsesLogger(model, logsDir = null) {
-  // Skip logging in worker environment (no fs)
-  if (typeof fs.mkdirSync !== "function") {
-    return null;
-  }
-
-  const timestamp = new Date().toISOString().replace(/[:.]/g, "").slice(0, 15);
-  const uniqueId = Math.random().toString(36).slice(2, 8);
-  const baseDir = logsDir || (typeof process !== "undefined" ? process.cwd() : ".");
-  // previous: const baseDir = logsDir || resolveDataDir(); — reverted in #555 for Workers compat
-  const logDir = path.join(baseDir, "logs", `responses_${model}_${timestamp}_${uniqueId}`);
-
-  try {
-    fs.mkdirSync(logDir, { recursive: true });
-  } catch {
-    return null;
-  }
-
-  let inputEvents = [];
-  let outputEvents = [];
-
-  return {
-    logInput: (event) => {
-      inputEvents.push(event);
-    },
-    logOutput: (event) => {
-      outputEvents.push(event);
-    },
-    flush: () => {
-      try {
-        fs.writeFileSync(path.join(logDir, "1_input_stream.txt"), inputEvents.join("\n"));
-        fs.writeFileSync(path.join(logDir, "2_output_stream.txt"), outputEvents.join("\n"));
-      } catch (e) {
-        console.log("[RESPONSES] Failed to write logs:", e.message);
-      }
-    },
-  };
-}
 
 /**
  * Create TransformStream that converts Chat Completions SSE to Responses API SSE
